refactor(#131): extract CLI structs and logging into dedicated modules
Extract CLI argument definitions (clap structs/enums) into cli.rs and
logging setup into logging.rs. main.rs reduced from 449 to ~100 lines.

- cli.rs: Cli struct, Commands enum, TagCommands, NamespaceCommands,
  AgingCommands, SupportedShell
- logging.rs: init() with console + daily rotating file appender
- main.rs: entry point only (parse, logging, config, dispatch)
- All crate::X imports updated to crate::cli::X in command modules
- logging.rs: create log dir if missing, fallback to temp instead of cwd
- Added .cora.yaml config (pre-commit hook installed)

Also adds .cora/ to .gitignore (local Cora cache directory).
