+++ .cora.yaml
@@ -1,0 +1,41 @@
+# Cora Code Review Configuration
+# See: https://github.com/codecoradev/cora-cli
+
+# Provider settings — matches CI (CORA_API_KEY, CORA_BASE_URL, CORA_MODEL)
+provider:
+  provider: freemodel
+  model: gpt-5.4
+  base_url: https://api.freemodel.dev/v1
+
+# Review focus areas: security, performance, bugs, best-practice, style
+focus:
+  - security
+  - performance
+  - bugs
+
+# Custom rules / additional instructions for the reviewer
+rules:
+  - "All SQL queries use parameterized/bound statements via rusqlite. CLI args are struct fields passed to uteke-core which uses param() bindings, never string interpolation. Do not flag SQL injection on CLI struct definitions."
+
+# Ignore patterns
+ignore:
+  files:
+    - "*.lock"
+    - "package-lock.json"
+    - "yarn.lock"
+    - "vendor/"
+    - "node_modules/"
+  rules: []
+#   rules:
+#     - "style"  # ignore style issues
+
+# Hook settings
+hook:
+  mode: warn          # "warn" (print but allow) or "block" (fail commit)
+  min_severity: major # minimum severity to trigger hook action
+  max_diff_size: 5242880 # max 5MB diff chars before refusing
+
+# Output settings
+output:
+  format: pretty      # pretty, json, compact, sarif
+  color: true
