+++ .gitignore
@@ -14,3 +14,4 @@ uteke_index.keys
 uteke_index.usearch
 .fusion/
 progress.md
+.cora/
